refactor(reddit): nest Reddit into folders by sub-namespace
Reorganise RedditPodcastPoster.Reddit so folder structure mirrors
namespace, matching the Apple/Abstractions/Spotify/YouTube layout
convention (was ~96% root-level files, 27 of 28 .cs files).

Introduce Clients/, Configuration/, Factories/, Managers/, Models/,
and Posters/ folders, each 1:1 with a matching namespace under
RedditPodcastPoster.Reddit; the existing Extensions/ folder is
untouched. Public type names and the AddRedditServices()/
AddSubredditSettings()/AddRedditClient()/AddAdminRedditClient() DI
extension method names are unchanged. No behaviour change.

Update using statements across ~8 consumer files solution-wide
(Cloud/Api, Common, ContentPublisher, Subreddit, and the Reddit test
project) to reference the new sub-namespaces.
