+++ Class-Libraries/RedditPodcastPoster.Reddit.Tests/RedditEpisodeCommentFactoryTests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using Moq.AutoMock;
 using RedditPodcastPoster.Models;
+using RedditPodcastPoster.Reddit.Factories;
 
 namespace RedditPodcastPoster.Reddit.Tests;
