+++ Class-Libraries/RedditPodcastPoster.Reddit/Clients/IAdminRedditClient.cs
@@ -1,6 +1,6 @@
 ﻿using Reddit;
 
-namespace RedditPodcastPoster.Reddit;
+namespace RedditPodcastPoster.Reddit.Clients;
 
 public interface IAdminRedditClient
 {
