+++ Class-Libraries/RedditPodcastPoster.Reddit/Configuration/RedditSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace RedditPodcastPoster.Reddit;
+﻿namespace RedditPodcastPoster.Reddit.Configuration;
 
 public class RedditSettings
 {
