+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IAdminRedditClientFactory.cs
@@ -1,0 +1,8 @@
+﻿using RedditPodcastPoster.Reddit.Clients;
+
+namespace RedditPodcastPoster.Reddit.Factories;
+
+public interface IAdminRedditClientFactory
+{
+    IAdminRedditClient Create();
+}
