+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IRedditBundleCommentFactory.cs
@@ -1,3 +1,3 @@
-﻿namespace RedditPodcastPoster.Reddit;
+﻿namespace RedditPodcastPoster.Reddit.Factories;
 
 public interface IRedditBundleCommentFactory : IRedditCommentFactory { }
