+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IRedditClientFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Reddit;
 
-namespace RedditPodcastPoster.Reddit;
+namespace RedditPodcastPoster.Reddit.Factories;
 
 public interface IRedditClientFactory
 {
