+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IRedditCommentFactory.cs
@@ -1,6 +1,6 @@
 ﻿using RedditPodcastPoster.Models;
 
-namespace RedditPodcastPoster.Reddit;
+namespace RedditPodcastPoster.Reddit.Factories;
 
 public interface IRedditCommentFactory
 {
