+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IRedditEpisodeCommentFactory.cs
@@ -1,3 +1,3 @@
-﻿namespace RedditPodcastPoster.Reddit;
+﻿namespace RedditPodcastPoster.Reddit.Factories;
 
 public interface IRedditEpisodeCommentFactory : IRedditCommentFactory { }
