+++ Class-Libraries/RedditPodcastPoster.Reddit/Factories/IRedditPostTitleFactory.cs
@@ -1,6 +1,6 @@
 ﻿using RedditPodcastPoster.Models;
 
-namespace RedditPodcastPoster.Reddit;
+namespace RedditPodcastPoster.Reddit.Factories;
 
 public interface IRedditPostTitleFactory
 {
