--- Class-Libraries/RedditPodcastPoster.Reddit/IAdminRedditClientFactory.cs
@@ -1,6 +1,0 @@
-﻿namespace RedditPodcastPoster.Reddit;
-
-public interface IAdminRedditClientFactory
-{
-    IAdminRedditClient Create();
-}
