+++ Class-Libraries/RedditPodcastPoster.Reddit/Resolvers/PostResolver.cs
@@ -3,8 +3,9 @@ using Microsoft.Extensions.Options;
 using Reddit;
 using Reddit.Controllers;
 using RedditPodcastPoster.Models;
+using RedditPodcastPoster.Reddit.Configuration;
 
-namespace RedditPodcastPoster.Reddit;
+namespace RedditPodcastPoster.Reddit.Resolvers;
 
 public class PostResolver(
     RedditClient redditClient,
