+++ Cloud/Api/Handlers/EpisodeHandler.cs
@@ -30,7 +30,7 @@ using RedditPodcastPoster.PodcastServices.Spotify.Resolvers;
 using RedditPodcastPoster.PodcastServices.YouTube;
 using RedditPodcastPoster.PodcastServices.YouTube.Extensions;
 using RedditPodcastPoster.PodcastServices.YouTube.Resolvers;
-using RedditPodcastPoster.Reddit;
+using RedditPodcastPoster.Reddit.Managers;
 using RedditPodcastPoster.People;
 using RedditPodcastPoster.People.Models;
 using RedditPodcastPoster.Subjects;
