+++ Cloud/Api/Handlers/SubjectHandler.cs
@@ -11,7 +11,8 @@ using RedditPodcastPoster.Auth0;
 using RedditPodcastPoster.ContentPublisher;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions;
-using RedditPodcastPoster.Reddit;
+using RedditPodcastPoster.Reddit.Clients;
+using RedditPodcastPoster.Reddit.Configuration;
 using RedditPodcastPoster.Subjects;
 using RedditPodcastPoster.Subjects.Factories;
 using Subject = RedditPodcastPoster.Models.Subject;
