+++ Cloud/Api/Ioc.cs
@@ -28,8 +28,8 @@ using RedditPodcastPoster.PodcastServices.Spotify.Extensions;
 using RedditPodcastPoster.PodcastServices.YouTube.Configuration;
 using RedditPodcastPoster.PodcastServices.YouTube.Extensions;
 using RedditPodcastPoster.PushSubscriptions.Extensions;
-using RedditPodcastPoster.Reddit;
 using RedditPodcastPoster.Reddit.Extensions;
+using RedditPodcastPoster.Reddit.Factories;
 using RedditPodcastPoster.People.Extensions;
 using RedditPodcastPoster.Subjects.Extensions;
 using RedditPodcastPoster.Text.Extensions;
